Add ability to reset pg_stat_slru in pg_stat_reset_shared()
Currently, pg_stat_reset_shared() can not reset pg_stat_slru
nevertheless this is one of shared statistics view.

This patch adds a new argment 'slru' to reset all the rows
in pg_stat_slru.

Considering pg_stat_reset_slru() was introduced in
28cac71bd368(PostgreSQL 13) and it could impact existing
applications and it can not only reset all the entry but
can specify which entry to reset, this function is left
as it is.